Conflicts:
	Scribus/scribus/CMakeLists.txt
	Scribus/scribus/notesstyles.cpp
	Scribus/scribus/notesstyles.h
	Scribus/scribus/pageitem.cpp
	Scribus/scribus/pageitem_noteframe.cpp
	Scribus/scribus/pageitem_noteframe.h
	Scribus/scribus/pageitem_textframe.cpp
	Scribus/scribus/scribusdoc.cpp
	Scribus/scribus/scribusdoc.h
	Scribus/scribus/text/storytext.cpp
	Scribus/scribus/text/storytext.h
	Scribus/scribus/util_text.cpp
	Scribus/scribus/util_text.h

+++ Scribus/scribus/collect4output.cpp
@@ -232,7 +232,7 @@ bool CollectForOutput::collectItems()
 				emit itemsCollected(c++);
 		}
 	}
-	for (QHash<int, PageItem*>::iterator itf = m_Doc->FrameItems.begin(); itf != m_Doc->FrameItems.end(); ++itf)
+	for (QHash<short, PageItem*>::iterator itf = m_Doc->FrameItems.begin(); itf != m_Doc->FrameItems.end(); ++itf)
 	{
 		PageItem *it = itf.value();
 		if (it->isGroup())
